Refactor coingecko types
