+++ src/common/interfaces/coingecko.interface.ts
@@ -1,0 +1,11 @@
+export interface Price {
+  timestamp: number;
+  value: number;
+}
+
+export interface PriceChart {
+  tokenId: string;
+  tokenName: string;
+  currentPriceUSD: number;
+  prices: Price[];
+}
