+++ src/common/interfaces/uniswap.interface.ts
@@ -1,0 +1,9 @@
+export interface Network {
+  id: string;
+  name: string;
+  desc: string;
+  logoURI: string;
+  disabled?: boolean;
+  isNew?: boolean;
+  subgraphEndpoint: string;
+}
