+++ src/common/network.ts
@@ -1,12 +1,5 @@
-export interface Network {
-  id: string;
-  name: string;
-  desc: string;
-  logoURI: string;
-  disabled?: boolean;
-  isNew?: boolean;
-  subgraphEndpoint: string;
-}
+import { Network } from "./interfaces/uniswap.interface";
+
 export const NETWORKS: Network[] = [
   {
     id: "ethereum",
