+++ src/context/app/appContext.tsx
@@ -1,6 +1,7 @@
 import React from "react";
-import { Network, NETWORKS } from "../../common/network";
-import { PriceChart } from "../../repos/coingecko";
+import { PriceChart } from "../../common/interfaces/coingecko.interface";
+import { Network } from "../../common/interfaces/uniswap.interface";
+import { NETWORKS } from "../../common/network";
 import { Pool, Tick, V3Token } from "../../repos/uniswap";
 import { AppAction, appReducer } from "./appReducer";
