+++ src/repos/coingecko.ts
@@ -1,17 +1,9 @@
 import axios from "axios";
+import { Price, PriceChart } from "../common/interfaces/coingecko.interface";
 import tokenAddressMapping from "./tokenAddressMapping.json";
 import { currentNetwork } from "./uniswap";
 
-export enum QueryPeriodEnum {
-  ONE_DAY = "1",
-  ONE_WEEK = "7",
-  ONE_MONTH = "30",
-  THREE_MONTH = "90",
-  ONE_YEAR = "90",
-  MAX = "max",
-}
-
-export interface Token {
+interface Token {
   id: string;
   name: string;
 }
@@ -30,15 +22,13 @@ export const getCoingeckoToken = (contractAddress: string): Token | null => {
   return null;
 };
 
-export interface Price {
-  timestamp: number;
-  value: number;
-}
-export interface PriceChart {
-  tokenId: string;
-  tokenName: string;
-  currentPriceUSD: number;
-  prices: Price[];
+enum QueryPeriodEnum {
+  ONE_DAY = "1",
+  ONE_WEEK = "7",
+  ONE_MONTH = "30",
+  THREE_MONTH = "90",
+  ONE_YEAR = "90",
+  MAX = "max",
 }
 export const getPriceChart = async (
   contractAddress: string,
